Got the lua packages working after figuring out Lua path
Has to be in a Lua subdirectory apparently. Nearly did that by accident - but I got to do 3 minutes figuring out it was LUA_PATH and then an hour figuring out how to set it for Alacritty, so who's the real winner? In the end gave up and added lua subfolder.